requirements.txt
textual[syntax]
python-frontmatter
markdown
jinja2
toml
Pillow

